refactor: consolidate visualization functions into train_task_helpers and remove unused visualization module
